test: Name test files for what they test, not review rounds
test_round5_fixes.py and test_round5_regressions.py were named after the
review round that produced them, which tells a later reader nothing, and
each mixed tests for five different modules. Split by module under test:

  test_action_filter_hardening.py    redaction + consumer-callback containment
  test_tempdir_hardening.py          shared temp root validation + cleanup
  test_generated_shell_script.py     POSIX action script quoting
  test_optimized_mode_invariants.py  invariants that must survive python -O
  test_subprocess_process_group.py   process-group recording
  test_runner_launch_state.py        runner state after a failed launch

Class names lose their finding IDs for the behaviour they pin, e.g.
TestR51RedactionArgsBypass -> TestRedactionDoesNotLeakViaRecordArgs.

Test bodies are unchanged: 76 tests before, 76 after, all passing. Full
suite still 855 passed / 39 skipped / 16 xfailed.
